src/services/astrology-python/requirements.txt
# Swiss Ephemeris for astronomical calculations
pyswisseph==2.10.3.2

# Kerykeion for chart generation
kerykeion==4.2.0

# Geographic services
geopy==2.4.1
timezonefinder==6.2.0

# Timezone handling
pytz>=2022.7,<2023.0

# Supabase client
supabase==2.3.5

# Additional dependencies
python-dateutil==2.8.2
requests==2.31.0
numpy==1.26.3